api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Pillow==10.1.0
requests==2.32.3
Werkzeug==3.0.1 